+++ packages/layerchart/src/routes/docs/components/Axis/+page.svelte
@@ -11,7 +11,6 @@
     timeMillisecond,
   } from 'd3-time';
   import { Field, RangeField, Switch, ToggleGroup, ToggleOption } from 'svelte-ux';
-  import { format, PeriodType } from '@layerstack/utils';
 
   import { Axis, Chart, Frame, Layer, Rule, Text, Grid, asAny } from 'layerchart';
   import Preview from '$lib/docs/Preview.svelte';
@@ -329,7 +328,7 @@
           placement="bottom"
           rule
           ticks={(scale) => scale.domain()}
-          format={(d) => format(d, PeriodType.Day, { variant: 'long' })}
+          format={{ type: 'day', options: { variant: 'long' } }}
         >
           {#snippet tickLabel({ props, index })}
             <Text {...props} textAnchor={index === 0 ? 'start' : 'end'} />
@@ -675,6 +674,7 @@
     label="tickSpacing"
     labelPlacement="left"
     bind:value={tickSpacing}
+    min={10}
     max={300}
     step={10}
   />
@@ -707,6 +707,7 @@
     label="tickSpacing"
     labelPlacement="left"
     bind:value={tickSpacing}
+    min={10}
     max={300}
     step={10}
   />
@@ -798,6 +799,7 @@
     label="tickSpacing"
     labelPlacement="left"
     bind:value={tickSpacing}
+    min={10}
     max={300}
     step={10}
   />
@@ -852,6 +854,7 @@
     label="tickSpacing"
     labelPlacement="left"
     bind:value={tickSpacing}
+    min={10}
     max={300}
     step={10}
   />