+++ packages/layerchart/src/routes/docs/examples/Beeswarm/+page.svelte
@@ -1,19 +1,26 @@
 <script lang="ts">
   import { scaleOrdinal } from 'd3-scale';
-  import { forceX, forceY, forceCollide } from 'd3-force';
-  import { PeriodType } from '@layerstack/utils';
+  import { forceX, forceY, forceCollide, type SimulationNodeDatum } from 'd3-force';
 
   import { asAny, Axis, Chart, Circle, ForceSimulation, Svg, Tooltip } from 'layerchart';
 
   import Preview from '$lib/docs/Preview.svelte';
+  import type { USSenatorsDatum } from '$static/data/examples/us-senators.js';
+  import type { Prettify } from '@layerstack/utils';
+
+  type NodeDatum = USSenatorsDatum;
+
+  type MySimulationNodeDatum = Prettify<NodeDatum & SimulationNodeDatum>;
 
   let { data } = $props();
 
+  const nodes: MySimulationNodeDatum[] = data.usSenators;
+
   const genderColor = scaleOrdinal(['var(--color-info)', 'var(--color-warning)']);
 
-  const xForce = forceX().strength(0.95);
-  const yForce = forceY().strength(0.075);
-  const collideForce = forceCollide();
+  const xForce = forceX<MySimulationNodeDatum>().strength(0.95);
+  const yForce = forceY<MySimulationNodeDatum>().strength(0.075);
+  const collideForce = forceCollide<MySimulationNodeDatum>();
 </script>
 
 <h1>Examples</h1>
@@ -36,6 +43,7 @@
               y: yForce.y(context.height / 2),
               collide: collideForce.radius(r),
             }}
+            data={{ nodes }}
             static
           >
             {#snippet children({ nodes })}
@@ -58,7 +66,7 @@
           {#snippet children({ data })}
             <Tooltip.Header>{data.name}</Tooltip.Header>
             <Tooltip.List>
-              <Tooltip.Item label="Birth date" value={data.date_of_birth} format={PeriodType.Day} />
+              <Tooltip.Item label="Birth date" value={data.date_of_birth} format="day" />
               <Tooltip.Item label="State" value={data.state_name} />
               <Tooltip.Item label="Party" value={data.party} />
               <Tooltip.Item label="Gender" value={data.gender} />