+++ packages/layerchart/src/routes/docs/examples/Line/+page.svelte
@@ -2,8 +2,6 @@
   import { scaleOrdinal, scaleSequential, scaleTime } from 'd3-scale';
   import { extent, flatGroup, ticks } from 'd3-array';
   import { interpolateTurbo } from 'd3-scale-chromatic';
-  import { format } from 'date-fns';
-  import { formatDate, PeriodType } from '@layerstack/utils';
   import { cls } from '@layerstack/tailwind';
 
   import {
@@ -129,7 +127,7 @@
 
       <Tooltip.Root>
         {#snippet children({ data })}
-          <Tooltip.Header>{format(data.date, 'eee, MMMM do')}</Tooltip.Header>
+          <Tooltip.Header value={data.date} format="day" />
           <Tooltip.List>
             <Tooltip.Item label="value" value={data.value} />
           </Tooltip.List>
@@ -273,7 +271,7 @@
 
         <Tooltip.Root>
           {#snippet children({ data })}
-            <Tooltip.Header>{format(data.date, 'eee, MMMM do')}</Tooltip.Header>
+            <Tooltip.Header value={data.date} format="day" />
             <Tooltip.List>
               <Tooltip.Item label={data.fruit} value={data.value} />
             </Tooltip.List>
@@ -367,7 +365,7 @@
           <Highlight points lines />
         </Svg>
         <Tooltip.Root>
-          <Tooltip.Header>{format(context.tooltip.data.date, 'eee, MMMM do')}</Tooltip.Header>
+          <Tooltip.Header value={context.tooltip.data.date} format="day" />
           <Tooltip.List>
             <Tooltip.Item label={context.tooltip.data.fruit} value={context.tooltip.data.value} />
           </Tooltip.List>
@@ -408,7 +406,7 @@
         </Svg>
         <Tooltip.Root>
           {#snippet children({ data })}
-            <Tooltip.Header>{format(data.date, 'eee, MMMM do')}</Tooltip.Header>
+            <Tooltip.Header value={data.date} format="day" />
             <Tooltip.List>
               <Tooltip.Item label={data.fruit} value={data.value} />
             </Tooltip.List>